Update filepaths for collaborative optimizations in QAT
