+++ tensorflow_model_optimization/python/core/api/BUILD
@@ -14,7 +14,7 @@ py_strict_library(
         "experimental/combine/__init__.py",
         "quantization/__init__.py",
         "quantization/keras/__init__.py",
-        "quantization/keras/collaborative_optimizations/__init__.py",
+        "quantization/keras/collab_opts/__init__.py",
         "quantization/keras/default_8bit/__init__.py",
         "quantization/keras/default_8bit/default_8bit_transforms/__init__.py",
         "quantization/keras/experimental/__init__.py",
@@ -42,9 +42,9 @@ py_strict_library(
         "//tensorflow_model_optimization/python/core/quantization/keras:quantize_scheme",
         "//tensorflow_model_optimization/python/core/quantization/keras:quantize_wrapper",
         "//tensorflow_model_optimization/python/core/quantization/keras:quantizers",
-        "//tensorflow_model_optimization/python/core/quantization/keras/collaborative_optimizations/cluster_preserve:cluster_utils",
-        "//tensorflow_model_optimization/python/core/quantization/keras/collaborative_optimizations/cluster_preserve:default_8bit_cluster_preserve_quantize_scheme",
-        "//tensorflow_model_optimization/python/core/quantization/keras/collaborative_optimizations/prune_preserve:default_8bit_prune_preserve_quantize_scheme",
+        "//tensorflow_model_optimization/python/core/quantization/keras/collab_opts/cluster_preserve:cluster_utils",
+        "//tensorflow_model_optimization/python/core/quantization/keras/collab_opts/cluster_preserve:default_8bit_cluster_preserve_quantize_scheme",
+        "//tensorflow_model_optimization/python/core/quantization/keras/collab_opts/prune_preserve:default_8bit_prune_preserve_quantize_scheme",
         "//tensorflow_model_optimization/python/core/quantization/keras/default_8bit:default_8bit_quantize_layout_transform",
         "//tensorflow_model_optimization/python/core/quantization/keras/default_8bit:default_8bit_quantize_registry",
         "//tensorflow_model_optimization/python/core/quantization/keras/default_8bit:default_8bit_quantize_scheme",
