+++ tensorflow_model_optimization/python/core/api/experimental/combine/__init__.py
@@ -14,13 +14,13 @@
 # ==============================================================================
 """Module containing collaborative optimization code."""
 
-from tensorflow_model_optimization.python.core.quantization.keras.collaborative_optimizations.cluster_preserve.cluster_utils import (
+from tensorflow_model_optimization.python.core.quantization.keras.collab_opts.cluster_preserve.cluster_utils import (
     strip_clustering_cqat,)
 
-from tensorflow_model_optimization.python.core.quantization.keras.collaborative_optimizations.cluster_preserve.default_8bit_cluster_preserve_quantize_scheme import (
+from tensorflow_model_optimization.python.core.quantization.keras.collab_opts.cluster_preserve.default_8bit_cluster_preserve_quantize_scheme import (
     Default8BitClusterPreserveQuantizeScheme,)
 
 # Deprecated import.
-# Please import from tfmot.quantization.keras.collaborative_optimizations
-from tensorflow_model_optimization.python.core.quantization.keras.collaborative_optimizations.prune_preserve.default_8bit_prune_preserve_quantize_scheme import (
+# Please import from tfmot.quantization.keras.collab_opts
+from tensorflow_model_optimization.python.core.quantization.keras.collab_opts.prune_preserve.default_8bit_prune_preserve_quantize_scheme import (
     Default8BitPrunePreserveQuantizeScheme,)
