+++ tensorflow_model_optimization/python/core/api/quantization/keras/__init__.py
@@ -19,7 +19,7 @@
 from tensorflow_model_optimization.python.core.api.quantization.keras import quantizers
 from tensorflow_model_optimization.python.core.api.quantization.keras import default_8bit
 from tensorflow_model_optimization.python.core.api.quantization.keras import graph_transformations
-from tensorflow_model_optimization.python.core.api.quantization.keras import collaborative_optimizations
+from tensorflow_model_optimization.python.core.api.quantization.keras import collab_opts
 from tensorflow_model_optimization.python.core.api.quantization.keras import experimental
 
 # quantize all layers with default quantization implementation.
