+++ tensorflow_model_optimization/python/examples/quantization_with_sparsity/keras/mnist_cnn.py
@@ -25,7 +25,7 @@ import numpy as np
 import tensorflow as tf
 
 from tensorflow_model_optimization.python.core.quantization.keras import quantize
-from tensorflow_model_optimization.python.core.quantization.keras.collaborative_optimizations.prune_preserve import (
+from tensorflow_model_optimization.python.core.quantization.keras.collab_opts.prune_preserve import (
     default_8bit_prune_preserve_quantize_scheme,)
 from tensorflow_model_optimization.python.core.sparsity.keras import prune
 from tensorflow_model_optimization.python.core.sparsity.keras import pruning_callbacks
